Telegram API connection header
### Explanation

This line of code is used to set the HTTP response header to inform the client that the content type of the response is JSON (JavaScript Object Notation).

#### Components of the Code

1. `header` function:
   - The header function in PHP is used to send raw HTTP headers to the client.
   - Syntax: header(string $header, bool $replace = true, int $http_response_code = 0)

2. Header String:
   - The string 'Content-Type: application/json' is the actual header being sent.
   - Content-Type is an HTTP header used to indicate the media type of the resource.
   - application/json specifies that the content is in JSON format.

### Purpose

By setting this header, the server is instructing the client (e.g., a web browser, an API client) that the data being sent is in JSON format. This is important for clients to correctly interpret and handle the response data.

### Usage Example

This is commonly used in APIs and web services where the server sends JSON responses. For example, if you have an API endpoint that returns user data in JSON format, you would set this header so that the client knows to parse the response as JSON.

#### Example in a PHP script:

<?php
header('Content-Type: application/json');

$response = array(
    'status' => 'success',
    'message' => 'Data retrieved successfully',
    'data' => array('id' => 1, 'name' => 'John Doe')
);

echo json_encode($response);
?>

In this example:
- The header('Content-Type: application/json'); line sets the content type of the response to JSON.
- An array $response is created with some data.
- json_encode($response) converts the PHP array into a JSON string.
- echo outputs the JSON string, which is then sent to the client.

This ensures the client correctly interprets the response as JSON.