PAY button
Explanation of Each Part
Creating an Array for the Keyboard Configuration:

The code creates a PHP array that represents the structure and configuration of a custom keyboard for a Telegram bot.
This array is then JSON-encoded to be sent to Telegram's API.
keyboard Key:

'keyboard' => [        
    [
        ['text' => "PAY"]
    ], 
],
The keyboard key contains an array of rows, each of which contains an array of buttons.
In this case, there is one row with one button.
Each button is represented as an array with a text key. Here, the button displays the text "PAY".
resize_keyboard Key:


"resize_keyboard" => true,
The resize_keyboard key is a boolean option.
If set to true, the custom keyboard will be resized to fit the number of buttons. Otherwise, it retains its default size.
one_time_keyboard Key:


'one_time_keyboard' => true
The one_time_keyboard key is also a boolean option.
If set to true, the custom keyboard will be hidden after the user makes a selection. If set to false, the keyboard remains on the screen.
json_encode Function:


json_encode([...])
The json_encode function converts the PHP array into a JSON string.
This JSON string is assigned to the variable $startKey.
Resulting JSON Structure
The resulting JSON string in $startKey would look like this:

{
    "keyboard": [
        [
            {
                "text": "PAY"
            }
        ]
    ],
    "resize_keyboard": true,
    "one_time_keyboard": true
}
Use Case
This JSON string can be used to send a custom keyboard to a user via Telegram's Bot API. For example, you would send this JSON in a request to the sendMessage method with the reply_markup field:


$chat_id = 'CHAT_ID_HERE'; // Replace with the actual chat ID
$message = 'Please choose an option:';
$reply_markup = $startKey;

$response = file_get_contents("https://api.telegram.org/botYOUR_BOT_TOKEN/sendMessage?chat_id=$chat_id&text=" . urlencode($message) . "&reply_markup=" . urlencode($reply_markup));
In this example:

YOUR_BOT_TOKEN should be replaced with your actual bot token.
CHAT_ID_HERE should be replaced with the actual chat ID to which you want to send the message.
The file_get_contents function sends a GET request to the Telegram API to send the message with the custom keyboard.
This approach is often used in Telegram bots to create interactive user interfaces where users can easily select predefined options.