BOT ID
This line of code is simple. It declares a variable called "$bot_id" and assigns it the integer value "".
In Telegram, go to BotFather (https://t.me/BotFather).
Send the /start command.
If you have already created your own bot, submit the /mybots command.
A list of your bots will be displayed. Choose the robot you want.
Then select the API Token option to get your bot's API token.
Your bot ID is contained in the API token information and looks like this:
123456:ABC-DEF1234ghIkl-zyx57W2v1u123ew11.
The number before the two digits is the ID of your Telegram bot